requirements-dev.txt
ipython     #terminal
ipdb        #debugger
sdb         #debugger remoto
pip-tools
pytest
pytest-order
httpx
black
flake8

fastapi[all]
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
httpx
pytest
python-dotenv
alembic
